Changed newgrasping into grasp
grasp is function that takes as input the coordinate of center of table.
It returns two values : 1.  'Picked' : means that the cylinder has been found and gripped ! It means that it is ready to put the object on a given basket
2. 'Unpicked' : No cylinder has been found

Before calling this function, the youbot has to face the table in parallel. (according to -y axis).